Shape: give each shape a name
The ShapeSorter is supposed to print a list of the shapes, sorted or
filtered as appropriate. This isn't very meaningful if the shapes aren't
called anything.

Give each shape a name by which it will be referred to.
Then the ShapeSorter can print a list of shapes' names and the output
will be more meaningful.